+++ lib/src/common/widgets/builder/column_builder.dart
@@ -28,6 +28,7 @@ class ColumnBuilder extends StatelessWidget {
         mainAxisSize: mainAxisSize,
         mainAxisAlignment: mainAxisAlignment,
         verticalDirection: verticalDirection,
-        children: List.generate(itemCount, (index) => itemBuilder(context, index)),
+        children:
+            List.generate(itemCount, (index) => itemBuilder(context, index)),
       );
 }