+++ lib/src/features/snapshot/feedback_plus.dart
@@ -2,8 +2,7 @@
 // It allows the user to give interactive feedback directly in the app.
 
 export 'src/better_feedback.dart';
-export 'src/feedback_controller.dart';
-export 'src/feedback_mode.dart';
-export 'src/l18n/translation.dart';
+export 'src/controllers/feedback_controller.dart';
+export 'src/models/user_feedback.dart';
 export 'src/theme/feedback_theme.dart' show FeedbackThemeData;
-export 'src/user_feedback.dart';
+export 'src/utilities/feedback_mode.dart';