+++ lib/src/features/snapshot/src/controllers/feedback_controller.dart
@@ -32,5 +32,6 @@ class FeedbackController extends ChangeNotifier {
   ///
   /// The controller is only attached if [FeedbackThemeData.sheetIsDraggable] is
   /// true and feedback is currently displayed.
-  final DraggableScrollableController sheetController = DraggableScrollableController();
+  final DraggableScrollableController sheetController =
+      DraggableScrollableController();
 }