+++ lib/src/features/snapshot/src/utilities/back_button_interceptor.dart
@@ -12,7 +12,8 @@ class BackButtonInterceptor with WidgetsBindingObserver {
   @visibleForTesting
   static const BackButtonInterceptor instance = BackButtonInterceptor._();
 
-  static final SplayTreeMap<double, List<BoolCallback>> _prioritizedCallbacks = SplayTreeMap();
+  static final SplayTreeMap<double, List<BoolCallback>> _prioritizedCallbacks =
+      SplayTreeMap();
 
   static void add(BoolCallback callback, {int? priority}) {
     if (_prioritizedCallbacks.isEmpty) {
@@ -22,14 +23,17 @@ class BackButtonInterceptor with WidgetsBindingObserver {
     if (callbacksList == null) {
       // Convert to double so that we have a valid maximum value to sort null
       // priorities last.
-      _prioritizedCallbacks[priority?.toDouble() ?? double.infinity] = [callback];
+      _prioritizedCallbacks[priority?.toDouble() ?? double.infinity] = [
+        callback,
+      ];
       return;
     }
     callbacksList.add(callback);
   }
 
   static void remove(BoolCallback callback) {
-    _prioritizedCallbacks.values.any((callbackList) => callbackList.remove(callback));
+    _prioritizedCallbacks.values
+        .any((callbackList) => callbackList.remove(callback));
     _prioritizedCallbacks.removeWhere((_, value) => value.isEmpty);
     if (_prioritizedCallbacks.isEmpty) {
       _unMount();
@@ -45,7 +49,8 @@ class BackButtonInterceptor with WidgetsBindingObserver {
     _binding.removeObserver(instance);
   }
 
-  Iterable<BoolCallback> get _callbacks => _prioritizedCallbacks.values.expand((lst) => lst);
+  Iterable<BoolCallback> get _callbacks =>
+      _prioritizedCallbacks.values.expand((lst) => lst);
 
   @override
   Future<bool> didPopRoute() async {