+++ lib/src/features/snapshot/src/utilities/renderer/renderer.dart
@@ -2,6 +2,7 @@
 // coverage:ignore-file
 
 import 'package:ispect/src/features/snapshot/src/utilities/renderer/_io_renderer.dart'
-    if (dart.library.html) 'package:ispect/src/features/snapshot/src/utilities/renderer/_html_renderer.dart' as impl;
+    if (dart.library.html) 'package:ispect/src/features/snapshot/src/utilities/renderer/_html_renderer.dart'
+    as impl;
 
 void printRendererErrorMessageIfNecessary() => impl.printErrorMessage();