+++ lib/src/features/snapshot/src/widgets/paint_on_background.dart
@@ -1,7 +1,7 @@
 // ignore_for_file: public_member_api_docs
 
 import 'package:flutter/material.dart';
-import 'package:ispect/src/features/snapshot/src/painter.dart';
+import 'package:ispect/src/features/snapshot/src/widgets/painter.dart';
 
 class PaintOnChild extends StatelessWidget {
   const PaintOnChild({